+++ lib/pruneInterfaces.d.ts
@@ -1,0 +1,3 @@
+import { DocumentNode, GraphQLResolveInfo } from 'graphql';
+declare const pruneLocalTypes: (doc: DocumentNode, prefix: string, info: GraphQLResolveInfo) => DocumentNode;
+export default pruneLocalTypes;