+++ lib/src/getDataLoader.d.ts
@@ -1,0 +1,3 @@
+import { EndpointDataLoader, ExecutionRef } from './types';
+declare const getDataLoader: (ref: ExecutionRef) => EndpointDataLoader<any>;
+export default getDataLoader;