+++ lib/src/nestGraphQLEndpoint.d.ts
@@ -1,0 +1,4 @@
+import { GraphQLSchema } from 'graphql';
+import { NestGraphQLEndpointParams } from './types';
+declare const nestGraphQLEndpoint: <TContext>(params: NestGraphQLEndpointParams<TContext>) => GraphQLSchema;
+export default nestGraphQLEndpoint;