+++ lib/src/wrapExecutor.d.ts
@@ -1,0 +1,3 @@
+import { Executor } from './types';
+declare const wrapExecutor: <TContext>(executor: Executor<TContext>, prefix: string) => Executor<TContext>;
+export default wrapExecutor;